fix: edge case bug for safe_name == original_name groups
When a group name is composed of an R valid name: e.g. 'sham' this results in adding parens to the name for each sample that shared the same group name.

This fix ensure the group name map is used once per unique group rather than once per sample for renaming purposes.